mood-recommender/backend/requirements.txt
# ─────────────────────────────────────────────────────────
#  Mood-Adaptive Recommender — Backend Requirements
# ─────────────────────────────────────────────────────────

# ── Web framework ─────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

# ── LLM ──────────────────────────────────────────────────
google-generativeai>=0.7.2

# ── Vector DB ────────────────────────────────────────────
faiss-cpu

# ── Embeddings ───────────────────────────────────────────
sentence-transformers==3.0.1
torch
numpy>=1.24.0

# ── HTTP client (for api_fetcher) ─────────────────────────
requests==2.32.3

# ── Environment ───────────────────────────────────────────
python-dotenv==1.0.1

# ── Streamlit frontend ────────────────────────────────────
streamlit==1.35.0

# ── Testing ───────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
